Ver.1.1.1

+++ app/controllers/sys/script/tasks_controller.rb
@@ -5,30 +5,28 @@ class Sys::Script::TasksController < ApplicationController
     tasks = task.find(:all, :order => :process_at)
     
     if tasks.size == 0
-      return render :text => "No Tasks"
+      return render(:text => "No Tasks")
     end
     
     tasks.each do |task|
       begin
         unless unid = task.unid_data
           task.destroy
-          raise 'Unid No Found'
+          raise 'Unid Not Found'
         end
         
         model = unid.model.underscore.pluralize
         item  = eval(unid.model).find_by_unid(unid.id)
         res   = render_component_as_string :controller => model.gsub(/^(.*?)\//, '\1/script/'),
           :action => task.name, :params => {:unid => unid, :item => item}
-        if res =~ /^OK/i
-          task.destroy
-        else
-          puts res
-        end
+        
+        puts res
+        task.destroy if res =~ /^OK/i
       rescue => e
-        puts "Error #{e}"
+        puts "Error: #{e}"
       end
     end
     
-    render :text => "OK"
+    render(:text => "OK")
   end
 end
